refactor(app): extract useZoomAndTheme hook
Phase 1 step 3 of the god-file cleanup plan. Move zoom + theme controls
out of App.tsx into src/hooks/useZoomAndTheme.ts.

The hook owns:
- applyZoom / adjustZoom / resetZoom (CSS zoom + --app-ui-scale token).
- setTheme (data-theme + sidebar.themes active flag).
- The window resize listener that re-syncs viewport vars at non-1.0 zoom.
- ZOOM_MIN / ZOOM_MAX constants.

App.tsx now calls `useZoomAndTheme({ setState, pushNotification })` and
destructures the action object. The boot-time persisted-zoom restore
remains inline because it's tangled with the broader config-load
sequence (font size, shell defaults, agent model, sidebar width) — that
moves with a future useConfig extraction.

App.tsx 5928 -> 5886 lines. tsc clean, eslint clean, 305/305 tests
pass. No behavior change.
